Better scanner start and remove circular dependency
