+++ backend/src/connectors/artefactScanning/wrapper.ts
@@ -20,31 +20,21 @@ export class ArtefactScanningWrapper {
     await Promise.all(
       Array.from(this.scanners).map(async (scanner) => {
         log.info({ toolName: scanner.toolName }, `Scanner initialising...`)
-        let attempt = 0
-        while (attempt <= config.connectors.artefactScanners.maxInitRetries) {
-          ++attempt
+        for (let attempt = 1; attempt <= config.connectors.artefactScanners.maxInitRetries; attempt++) {
           try {
-            await new Promise<void>((resolve, reject) => {
-              setTimeout(async () => {
-                try {
-                  await scanner.init()
-                } catch (error) {
-                  return reject(error)
-                }
-                log.info({ toolName: scanner.toolName }, `Scanner initialised`)
-                return resolve()
-              }, config.connectors.artefactScanners.initRetryDelay)
-            })
-            break
+            await scanner.init()
+            log.info({ toolName: scanner.toolName }, `Scanner initialised`)
+            return
           } catch (error) {
+            if (attempt === config.connectors.artefactScanners.maxInitRetries) {
+              throw ConfigurationError(
+                `Could not initialise scanner after max attempts, make sure that it is setup and configured correctly.`,
+                { failedAttempts: attempt, toolName: scanner.toolName, error },
+              )
+            }
             log.warn({ attempt: attempt, toolName: scanner.toolName, error }, `Could not initialise scanner, retrying.`)
+            await new Promise((res) => setTimeout(res, config.connectors.artefactScanners.initRetryDelay))
           }
-        }
-        if (attempt > config.connectors.artefactScanners.maxInitRetries) {
-          throw ConfigurationError(
-            `Could not initialise scanner after max attempts, make sure that it is setup and configured correctly.`,
-            { failedAttempts: attempt, toolName: scanner.toolName },
-          )
         }
       }),
     )
