+++ backend/test/connectors/artefactScanning/Base.spec.ts
@@ -1,8 +1,8 @@
 import PQueue from 'p-queue'
 import { beforeEach, describe, expect, test, vi } from 'vitest'
 
-import { ArtefactScanState, BaseArtefactScanningConnector } from '../../../src/connectors/artefactScanning/Base.js'
-import { ArtefactKind } from '../../../src/models/Scan.js'
+import { BaseArtefactScanningConnector } from '../../../src/connectors/artefactScanning/Base.js'
+import { ArtefactKind, ArtefactScanState } from '../../../src/models/Scan.js'
 
 vi.mock('../../../src/services/log.js')
